showTen() function working
showTen() is a function that displays 10 student-list elements for the
properly indexed page WHEN CLICKED.

Remaining issue is that function fails if page is not clicked, e.g.
when init() is called post search() call. Results are all returned
elements displayed on page 1, until another anchor element is clicked,
including anchor element for page 1.